refactor: Move first time modal to common
